fix: removed link variable from `bl_info` dictionary closes #1
In the dictionary `bl_info`, link values were placed as variable
because link was two characters too long. The line length in python
code should be 79 characters max and I was 81 characters in line.

This is a Official Python Documentation: PEP 8 -- Style Guide for
[Python Code](https://www.python.org/dev/peps/pep-0008/#id19).

This variable method on dictionary still worked in `Blender 2.93.1`.
Today I downloaded the latest `LTS` version
[Blender 2.93.2](https://www.blender.org/download/releases/2-93/)
and this method it doesn't work anymore.

So back to 81 characters in line.